requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
numpy==1.23.4
pika==1.3.1
pika-stubs==0.1.3
Pillow==9.3.0
python-dotenv==0.21.0
requests==2.28.1
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
typing_extensions==4.4.0
urllib3==1.26.12
